@@ -1,25 +1,31 @@
 {
   "name": "frontend",
   "version": "1.0.0",
-  "main": "index.js",
+  "description": "Speedboat Web Frontend",
+  "main": "serve.js",
   "license": "MIT",
+  "private": true,
+  "repository": {
+    "url": "https://github.com/dejaydev/speedboat",
+    "directory": "https://github.com/dejaydev/speedboat/tree/master/frontend"
+  },
+  "browserslist": [
+    "defaults"
+  ],
   "dependencies": {
-    "@babel/core": "^7.11.6",
-    "@babel/plugin-proposal-object-rest-spread": "^7.11.0",
+    "@babel/core": "^7.12.3",
     "@babel/plugin-transform-react-jsx": "^7.10.4",
-    "@babel/preset-env": "^7.11.5",
-    "@babel/preset-react": "^7.10.4",
-    "@babel/preset-stage-2": "^7.8.3",
+    "@babel/preset-env": "^7.12.1",
     "axios": "^0.20.0",
     "brace": "^0.11.1",
     "eventemitter3": "^4.0.7",
     "express": "^4.17.1",
-    "http-proxy-middleware": "^1.0.5",
+    "http-proxy-middleware": "^1.0.6",
     "lodash": "^4.17.20",
     "parcel-bundler": "^1.12.4",
-    "react": "^16.13.1",
-    "react-ace": "^9.1.3",
-    "react-dom": "^16.13.1",
+    "react": "^16.14.0",
+    "react-ace": "^9.1.4",
+    "react-dom": "^16.14.0",
     "react-router": "^5.2.0",
     "react-router-dom": "^5.2.0",
     "react-table": "^6.5.3",
