@@ -23,23 +23,17 @@ export default class Login extends Component {
     }
 
     return (
-      <div className="container">
-        <div className="row">
-          <div className="col-md-4 col-md-offset-4">
-            <div className="login-panel panel panel-default">
-              <div className="panel-heading">
-                <h3 className="panel-title">Login with Discord</h3>
-              </div>
-              <div className="panel-body">
-                <a href="/api/auth/discord">
-                  <img src="https://u.swvn.io/2017/OENkw.svg" height="256" width="256" style={{
-                    margin: 'auto',
-                    display: 'block',
-                  }} />
-                </a>
-              </div>
-            </div>
-          </div>
+      <div className="card">
+        <div className="card-header">
+          <h2 className="card-title text-primary text-center">Login with Discord</h2>
+        </div>
+        <div className="card-body">
+          <a href="/api/auth/discord">
+            <img src="https://dejay.dev/assets/discord.svg" height="256" width="256" style={{
+              margin: 'auto',
+              display: 'block',
+            }} />
+          </a>
         </div>
       </div>
     );
