@@ -2,12 +2,10 @@ import React, { Component } from 'react';
 
 class PageHeader extends Component {
   render() {
-		return (
-			<div className="row">
-				<div className="col-lg-12">
-					<h1 className="page-header">{this.props.name}</h1>
-				</div>
-			</div>
+	return (
+		<div className="d-sm-flex align-items-center justify-content-between mb-4">
+			<h1 className="h3 mb-0 text-primary">{this.props.name}</h1>
+		</div>
     );
   }
 }
