@@ -134,6 +134,8 @@ class Infraction(ModelBase):
     created_at = DateTimeField(default=datetime.utcnow)
     active = BooleanField(default=True)
 
+    messaged = BooleanField(default=False, null=True)
+
     class Meta:
         table_name = 'infractions'
 
@@ -151,7 +153,7 @@ class Infraction(ModelBase):
             'expires_at': self.expires_at,
             'created_at': self.created_at,
             'active': self.active,
-            'msg_status': self.msg_status,
+            'messaged': self.messaged,
         }
 
         base['type'] = {
@@ -241,7 +243,7 @@ class Infraction(ModelBase):
             actor_id=event.author.id,
             type_=cls.Types.KICK,
             reason=reason,
-            msg_status=msg_status)
+            messaged=msg_status)
 
     @classmethod
     def tempban(cls, plugin, event, member, reason, expires_at):
@@ -276,7 +278,7 @@ class Infraction(ModelBase):
             type_=cls.Types.TEMPBAN,
             reason=reason,
             expires_at=expires_at,
-            msg_status=msg_status)
+            messaged=msg_status)
 
     @classmethod
     def softban(cls, plugin, event, member, reason):
@@ -310,7 +312,7 @@ class Infraction(ModelBase):
             actor_id=event.author.id,
             type_=cls.Types.SOFTBAN,
             reason=reason,
-            msg_status=msg_status)
+            messaged=msg_status)
 
     @classmethod
     def ban(cls, plugin, event, member, reason, guild):
@@ -347,7 +349,7 @@ class Infraction(ModelBase):
             actor_id=event.author.id,
             type_=cls.Types.BAN,
             reason=reason,
-            msg_status=msg_status)
+            messaged=msg_status)
 
     @classmethod
     def warn(cls, plugin, event, member, reason, guild):
@@ -372,7 +374,7 @@ class Infraction(ModelBase):
             actor_id=event.author.id,
             type_=cls.Types.WARNING,
             reason=reason,
-            msg_status=msg_status)
+            messaged=msg_status)
 
     @classmethod
     def mute(cls, plugin, event, member, reason):
@@ -497,9 +499,9 @@ class Infraction(ModelBase):
             ))
                             
             msg_status = True
-        except APIException as err:
+        except Exception as err:
             msg_status = False  # Multiple bad things can happen here, so we'll just... do this.
-            print('Could not DM member {} because {}'.format(user, err))
+            print('{} - {}'.format(user, err.msg))
     
         return msg_status
 
